requirements.txt
music21==8.3.0
pygame==2.5.2
mido==1.3.2
pyfluidsynth==1.3.2
Pillow==10.0.0
